missing rule violation added, support for xaml (and other) file types
added "Sonar.UnknownIssueType" rule which is used to report IssueTypes
that were not known to this plugin, with some instructions to submit the
tags back to me for inclusion in future plugin release.
FIXES #1

when a violation is reported against a file that is not supported by the
dotnet plugins (non-cs or non-vb files, such as xaml files), log those
violations anyway.
FIXES #5