ability to supplement local rule via web UI property
You can now add IssueTypes to the 'ReSharper custom rules' property in
the Settings section of the web UI.  Added a note to the missing rules
violation to add the missing IssueTypes to the new property.
FIXES #2

modified the way the initial rules are populated to remove the need for
a seperate rules.xml file.